+++ codevr/Cargo.toml
@@ -5,10 +5,11 @@ authors = ["Alain Galvan <hi@alain.xyz>"]
 build = "build.rs"
 
 [dependencies]
+cgmath = "0.14.1"
 image = "0.13.0"
 libloading = "0.3.4"
 openvr = "0.1.1"
-serde = "1.0.6"
+serde = "1.0.1"
 serde_derive = "1.0.6"
 serde_json = "1.0.2"
 vulkano = "0.3.2"
