--- codevr/src/renderer/mod.rs
@@ -1,5 +1,0 @@
-mod text;
-
-struct Renderer {
-  
-}
